Update signed to unsigned

+++ src/types.ts
@@ -47,15 +47,15 @@ export enum EquipPosition {
 }
 
 export class Attire {
-  head: i32 = 0
-  neck: i32 = 0
-  body: i32 = 0
-  arms: i32 = 0
-  legs: i32 = 0
-  feet: i32 = 0
-  ring: i32 = 0
-  reserved1: i32 = 0
-  queueId: i64 = 0
+  head: u16 = 0
+  neck: u16 = 0
+  body: u16 = 0
+  arms: u16 = 0
+  legs: u16 = 0
+  feet: u16 = 0
+  ring: u16 = 0
+  reserved1: u16 = 0
+  queueId: u64 = 0
 }
 
 export enum ActionQueueStatus {
@@ -71,21 +71,21 @@ export enum CombatStyle {
 }
 
 export class Equipment {
-  itemTokenId: i32 = 0
-  amount: i32 = 0
+  itemTokenId: u16 = 0
+  amount: u16 = 0
 }
 
 export class QueuedActionInput {
   attire: Attire = new Attire()
-  actionId: i32 = 0
-  regenerateId: i32 = 0 // Food (combat), maybe something for non-combat later
-  choiceId: i32 = 0 // Melee/Arrow/Magic (combat), logs, ore (non-combat)
-  choiceId1: i32 = 0 // Reserved (TBD)
-  choiceId2: i32 = 0 // Reserved (TBD)
+  actionId: u32 = 0
+  regenerateId: u16 = 0 // Food (combat), maybe something for non-combat later
+  choiceId: u32 = 0 // Melee/Arrow/Magic (combat), logs, ore (non-combat)
+  choiceId1: u32 = 0 // Reserved (TBD)
+  choiceId2: u32 = 0 // Reserved (TBD)
   combatStyle: CombatStyle = CombatStyle.NONE
-  timespan: i32 = 0 // How long to queue the action for
-  rightHandEquipmentTokenId: i32 = 0
-  leftHandEquipmentTokenId: i32 = 0
+  timespan: u32 = 0 // How long to queue the action for
+  rightHandEquipmentTokenId: u16 = 0
+  leftHandEquipmentTokenId: u16 = 0
   skill: Skill = Skill.NONE
 }
 
@@ -94,27 +94,27 @@ export class ActionInfo {
   isAvailable: boolean = true
   isDynamic: boolean = false
   actionChoiceRequired: boolean = false
-  xpPerHour: i32 = 0
-  numSpawn: i32 = 0
-  minXP: i32 = 0
-  handItemTokenIdRangeMin: i32 = 0
-  handItemTokenIdRangeMax: i32 = 0
-  successPercent: i32 = 0
+  xpPerHour: u32 = 0
+  numSpawn: u32 = 0
+  minXP: u32 = 0
+  handItemTokenIdRangeMin: u16 = 0
+  handItemTokenIdRangeMax: u16 = 0
+  successPercent: u8 = 0
 }
 
 export class GuaranteedReward {
-  itemTokenId: i32 = 0
-  rate: i32 = 0 // base 100, 2 decimal places
+  itemTokenId: u16 = 0
+  rate: u32 = 0 // base 100, 2 decimal places
 }
 
 export class RandomReward {
   itemTokenId: i32 = 0
-  chance: i32 = 0 // out of 65335
-  amount: i32 = 0
+  chance: u16 = 0 // out of 65335
+  amount: u8 = 0
 }
 
 export class ActionInput {
-  actionId: i32 = 0
+  actionId: u32 = 0
   info: ActionInfo = new ActionInfo()
   guaranteedRewards: GuaranteedReward[] = []
   randomRewards: RandomReward[] = []
@@ -125,36 +125,36 @@ export class ActionInput {
 export class InputItem {
   combatStats: CombatStats = new CombatStats()
   nonCombatStats: NonCombatStats = new NonCombatStats()
-  tokenId: i32 = 0
+  tokenId: u16 = 0
   equipPosition: EquipPosition = EquipPosition.NONE
   // Can this be transferred to another player?
   isTransferable: boolean = true
   // Minimum requirements in this skill
   skill: Skill = Skill.NONE
-  minXP: i32 = 0
+  minXP: u32 = 0
   // Food
   healthRestored: i32 = 0
   // Boost
   boostType: BoostType = BoostType.NONE
-  boostValue: i32 = 0 // Varies, could be the % increase
-  boostDuration: i32 = 0 // How long the effect of the boost last
+  boostValue: u8 = 0 // Varies, could be the % increase
+  boostDuration: u32 = 0 // How long the effect of the boost last
   // uri
   metadataURI: string = ''
 }
 
 export class CombatStats {
-  melee: i32 = 0
-  magic: i32 = 0
-  range: i32 = 0
-  meleeDefence: i32 = 0
-  magicDefence: i32 = 0
-  rangeDefence: i32 = 0
-  health: i32 = 0
+  melee: i16 = 0
+  magic: i16 = 0
+  range: i16 = 0
+  meleeDefence: i16 = 0
+  magicDefence: i16 = 0
+  rangeDefence: i16 = 0
+  health: i16 = 0
 }
 
 export class NonCombatStats {
   skill: Skill = Skill.NONE
-  diff: i32 = 0
+  diff: u32 = 0
 }
 
 export const emptyCombatStats = new CombatStats()
@@ -199,57 +199,57 @@ export class Player {
   mintedNumber: string = '0' // Will be unique, use for sorting
   mintedTimestamp: string = '0'
   isActive: boolean = false // Is this player the active one for the owner
-  numActivities: i32 = 0 // Not used yet
+  numActivities: u32 = 0 // Not used yet
   pendingRandomRewards: string[] = [] // Timestamps for any rewards which are waiting on the next seed
-  speedMultiplier: i32 = 1
+  speedMultiplier: u32 = 1
 
   /* Action XP */
-  woodcuttingXP: i32 = 0
-  firemakingXP: i32 = 0
-  fishingXP: i32 = 0
-  cookingXP: i32 = 0
-  miningXP: i32 = 0
-  smithingXP: i32 = 0
-  craftingXP: i32 = 0
-  thievingXP: i32 = 0
+  woodcuttingXP: u32 = 0
+  firemakingXP: u32 = 0
+  fishingXP: u32 = 0
+  cookingXP: u32 = 0
+  miningXP: u32 = 0
+  smithingXP: u32 = 0
+  craftingXP: u32 = 0
+  thievingXP: u32 = 0
   totalXP: string = '0'
 
   /* Combat XP */
-  healthXP: i32 = 0
-  meleeXP: i32 = 0
-  defenceXP: i32 = 0
-  magicXP: i32 = 0
-  rangeXP: i32 = 0
+  healthXP: u32 = 0
+  meleeXP: u32 = 0
+  defenceXP: u32 = 0
+  magicXP: u32 = 0
+  rangeXP: u32 = 0
 
   /** Ranks */
-  woodcuttingRank: i32 = 0
-  firemakingRank: i32 = 0
-  fishingRank: i32 = 0
-  cookingRank: i32 = 0
-  miningRank: i32 = 0
-  smithingRank: i32 = 0
-  craftingRank: i32 = 0
-  thievingRank: i32 = 0
-  meleeRank: i32 = 0
-  magicRank: i32 = 0
-  rangeRank: i32 = 0
-  defenceRank: i32 = 0
-  healthRank: i32 = 0
-  combinedRank: i32 = 0
+  woodcuttingRank: u32 = 0
+  firemakingRank: u32 = 0
+  fishingRank: u32 = 0
+  cookingRank: u32 = 0
+  miningRank: u32 = 0
+  smithingRank: u32 = 0
+  craftingRank: u32 = 0
+  thievingRank: u32 = 0
+  meleeRank: u32 = 0
+  magicRank: u32 = 0
+  rangeRank: u32 = 0
+  defenceRank: u32 = 0
+  healthRank: u32 = 0
+  combinedRank: u32 = 0
 
   /* Boost */
-  boostStartTime: i32 = 0
-  boostDuration: i32 = 0
-  boostVal: i32 = 0
+  boostStartTime: u64 = 0
+  boostDuration: u32 = 0
+  boostVal: u8 = 0
   boostType: BoostType
-  boostItemTokenId: i32 = 0
+  boostItemTokenId: u16 = 0
 }
 
 export class User {
   id: string = '0'
   totalSold: string = '0'
   totalBought: string = '0'
-  numPlayers: i32 = 0
+  numPlayers: u32 = 0
 }
 
 export class GlobalPlayerStats {
@@ -272,7 +272,7 @@ export enum ActivityType {
 export class Activity {
   id: string = ''
   type: string = '' // ActivityType
-  itemTokenIds: i32[] = []
+  itemTokenIds: u16[] = []
   amounts: string[] = []
   prices: string[] = []
   queueId: string = ''
@@ -291,70 +291,70 @@ export class LastFullEquipment {
   id: string = '0' // playerId
   playerId: string = '0'
   user: string = ''
-  actionId: i32 = 0
-  regenerateId: i32 = 0
-  choiceId: i32 = 0
-  choiceId1: i32 = 0
-  choiceId2: i32 = 0
-  rightHandEquipmentTokenId: i32 = 0
-  leftHandEquipmentTokenId: i32 = 0
+  actionId: u32 = 0
+  regenerateId: u16 = 0
+  choiceId: u32 = 0
+  choiceId1: u32 = 0
+  choiceId2: u32 = 0
+  rightHandEquipmentTokenId: u16 = 0
+  leftHandEquipmentTokenId: u16 = 0
   startTime: string = ''
-  timespan: i32 = 0
+  timespan: u32 = 0
   isValid: boolean = true
   skill: Skill = Skill.NONE
 
-  headEquipped: i32 = 0
-  neckEquipped: i32 = 0
-  bodyEquipped: i32 = 0
-  armsEquipped: i32 = 0
-  legsEquipped: i32 = 0
-  feetEquipped: i32 = 0
-  ringEquipped: i32 = 0
+  headEquipped: u16 = 0
+  neckEquipped: u16 = 0
+  bodyEquipped: u16 = 0
+  armsEquipped: u16 = 0
+  legsEquipped: u16 = 0
+  feetEquipped: u16 = 0
+  ringEquipped: u16 = 0
   timestamp: string = '0'
 }
 
 export class QueuedAction {
   id: string = '' // queueId
-  actionId: i32 = 0
+  actionId: u32 = 0
   playerId: string = ''
-  regenerateId: i32 = 0
-  choiceId: i32 = 0
-  choiceId1: i32 = 0
-  choiceId2: i32 = 0
-  rightHandEquipmentTokenId: i32 = 0
-  leftHandEquipmentTokenId: i32 = 0
+  regenerateId: u16 = 0
+  choiceId: u32 = 0
+  choiceId1: u32 = 0
+  choiceId2: u32 = 0
+  rightHandEquipmentTokenId: u16 = 0
+  leftHandEquipmentTokenId: u16 = 0
   startTime: string = ''
-  timespan: i32 = 0
+  timespan: u32 = 0
   isValid: boolean = true
   skill: Skill = Skill.NONE
 
-  headEquipped: i32 = 0
-  neckEquipped: i32 = 0
-  bodyEquipped: i32 = 0
-  armsEquipped: i32 = 0
-  legsEquipped: i32 = 0
-  feetEquipped: i32 = 0
-  ringEquipped: i32 = 0
+  headEquipped: u16 = 0
+  neckEquipped: u16 = 0
+  bodyEquipped: u16 = 0
+  armsEquipped: u16 = 0
+  legsEquipped: u16 = 0
+  feetEquipped: u16 = 0
+  ringEquipped: u16 = 0
 }
 
 export class Item {
   id: string = '0'
-  tokenId: i32 = 0
-  melee: i32 = 0
-  magic: i32 = 0
-  range: i32 = 0
-  meleeDefence: i32 = 0
-  magicDefence: i32 = 0
-  rangeDefence: i32 = 0
-  health: i32 = 0
+  tokenId: u16 = 0
+  melee: i16 = 0
+  magic: i16 = 0
+  range: i16 = 0
+  meleeDefence: i16 = 0
+  magicDefence: i16 = 0
+  rangeDefence: i16 = 0
+  health: i16 = 0
   totalQuantity: string = '0'
-  equipPosition: i32 = 0
-  healthRestored: i32 = 0
+  equipPosition: u16 = 0
+  healthRestored: u32 = 0
   skill1: Skill = Skill.NONE
-  skillDiff1: i32 = 0
+  skillDiff1: u8 = 0
   boostType: BoostType = BoostType.NONE
-  boostValue: i32 = 0
-  boostDuration: i32 = 0
+  boostValue: u32 = 0
+  boostDuration: u32 = 0
   // Minimum requirements for equipping
   skill: Skill = Skill.NONE
   minXP: string = '0'
@@ -363,62 +363,62 @@ export class Item {
 export class UserItemNFT {
   id: string = ''
   user: string = ''
-  tokenId: i32 = 0
+  tokenId: u16 = 0
   amount: string = ''
   item: Item = new Item()
 }
 
 export class ShopItem {
   id: string = ''
-  tokenId: i32 = 0
+  tokenId: u16 = 0
   price: string = '0'
 }
 
 export class ActionChoice {
   id: string = ''
-  actionId: i32 = 0
+  actionId: u32 = 0
   skill: Skill = Skill.NONE
-  diff: i32 = 0
-  rate: i32 = 0
-  xpPerHour: i32 = 0
-  minXP: i32 = 0
-  inputTokenId1: i32 = 0
-  num1: i32 = 0
-  inputTokenId2: i32 = 0
-  num2: i32 = 0
-  inputTokenId3: i32 = 0
-  num3: i32 = 0
-  outputTokenId: i32 = 0
-  outputNum: i32 = 1
-  successPercent: i32 = 100
+  diff: u32 = 0
+  rate: u16 = 0
+  xpPerHour: u32 = 0
+  minXP: u32 = 0
+  inputTokenId1: u16 = 0
+  num1: u16 = 0
+  inputTokenId2: u16 = 0
+  num2: u16 = 0
+  inputTokenId3: u16 = 0
+  num3: u16 = 0
+  outputTokenId: u16 = 0
+  outputNum: u8 = 1
+  successPercent: u8 = 100
 }
 
 export class Action {
   id: string = ''
-  actionId: i32 = 0
+  actionId: u32 = 0
   skill: Skill = Skill.NONE
-  xpPerHour: i32 = 0
-  numSpawn: i32 = 0
+  xpPerHour: u32 = 0
+  numSpawn: u32 = 0
   minXP: string = '0'
 
-  guaranteedRewardItemTokenIds: i32[] = []
-  guaranteedRewardRates: i32[] = []
-  randomChanceItemTokenIds: i32[] = []
-  randomChanceRates: i32[] = []
+  guaranteedRewardItemTokenIds: u32[] = []
+  guaranteedRewardRates: u32[] = []
+  randomChanceItemTokenIds: u32[] = []
+  randomChanceRates: u32[] = []
 
-  handItemTokenIdRangeMin: i32 = 0
-  handItemTokenIdRangeMax: i32 = 0
+  handItemTokenIdRangeMin: u32 = 0
+  handItemTokenIdRangeMax: u32 = 0
   isAvailable: boolean = false
   isDynamic: boolean = false
   actionChoiceRequired: boolean = false
-  successPercent: i32 = 100
+  successPercent: i8 = 100
 
   /* Combat Stats */
-  melee: i32 = 0
-  magic: i32 = 0
-  range: i32 = 0
-  rangeDefence: i32 = 0
-  meleeDefence: i32 = 0
-  magicDefence: i32 = 0
-  health: i32 = 0
+  melee: i16 = 0
+  magic: i16 = 0
+  range: i16 = 0
+  rangeDefence: i16 = 0
+  meleeDefence: i16 = 0
+  magicDefence: i16 = 0
+  health: i16 = 0
 }
